move the test to test_lambda_sqs_integration.py
