+++ tests/integration/test_sqs.snapshot.json
@@ -357,44 +357,5 @@
       "create_queue_01": "An error occurred (QueueAlreadyExists) when calling the CreateQueue operation: A queue already exists with the same name and a different value for attribute DelaySeconds",
       "create_queue_02": "An error occurred (QueueAlreadyExists) when calling the CreateQueue operation: A queue already exists with the same name and a different value for attribute DelaySeconds"
     }
-  },
-  "tests/integration/test_sqs.py::TestSqsQueryApi::test_sqs_queue_as_lambda_dead_letter_queue": {
-    "recorded-date": "09-08-2022, 13:39:00",
-    "recorded-content": {
-      "lambda-response-dlq-config": {
-        "TargetArn": "arn:aws:sqs:<region>:111111111111:<resource:1>"
-      },
-      "messages": {
-        "Messages": [
-          {
-            "Body": {
-              "raise_error": 1
-            },
-            "MD5OfBody": "ba08269cbd0ce515889bfba6593731e1",
-            "MD5OfMessageAttributes": "<md5-hash>",
-            "MessageAttributes": {
-              "ErrorCode": {
-                "DataType": "Number",
-                "StringValue": "200"
-              },
-              "ErrorMessage": {
-                "DataType": "String",
-                "StringValue": "Test exception (this is intentional)"
-              },
-              "RequestID": {
-                "DataType": "String",
-                "StringValue": "<request-id:1>"
-              }
-            },
-            "MessageId": "<uuid:1>",
-            "ReceiptHandle": "<receipt-handle:1>"
-          }
-        ],
-        "ResponseMetadata": {
-          "HTTPHeaders": {},
-          "HTTPStatusCode": 200
-        }
-      }
-    }
   }
 }
